+++ config/filesystems.php
@@ -55,6 +55,13 @@ return [
             'visibility' => 'public',
         ],
 
+        'screenshots' => [
+            'driver' => 'local',
+            'root' => storage_path('app/public/screenshots'),
+            'url' => env('APP_URL').'/storage/screenshots',
+            'visibility' => 'public',
+        ],
+
         's3' => [
             'driver' => 's3',
             'key' => env('AWS_ACCESS_KEY_ID'),